requirements.txt
asgiref==3.4.1
Django==3.2.7
django-environ==0.7.0
django-filter==21.1
djangorestframework==3.12.4
environ==1.0
pytz==2021.1
sqlparse==0.4.2
utils==1.0.1
